backend/requirements.txt
google-genai
google-adk
fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx
aiofiles

# Testing dependencies
pytest
pytest-asyncio
pytest-cov
pytest-mock
httpx
